+++ packages/cli/src/cli.ts
@@ -1,7 +1,8 @@
 import process from 'node:process';
 import debug from 'debug';
+import updateNotifier, { type Package } from 'update-notifier';
 import minimist from 'minimist';
-import { createNew, serve } from './commands/index.js';
+import { convert, createNew, link, serve } from './commands/index.js';
 import { getPackageJson } from './util.js';
 
 const help = `Usage: moaw <command> [options]
@@ -12,6 +13,12 @@ Commands:
     -p, --port       Port to listen on (default: 4444)
     -h, --host       Host address to bind to (default: localhost)
     -o, --open       Open in browser (default: false)
+  c, convert <file>  Convert asciidoc to markdown
+    -a, --attr <json_file>  Attributes to use for conversion
+    -d, --dest <file>       Destination file (default: workshop.md)
+  l, link [<file>]   Get link to target file (default: workshop.md)
+    -r, --repo       Set GitHub repo instead of fetching it from git
+    -b, --branch <name>     Set branch name (default: current branch)
 
 General options:
   -v, --version      Show version
@@ -20,18 +27,24 @@ General options:
 
 export async function run(args: string[]) {
   const options = minimist(args, {
-    string: ['host'],
+    string: ['host', 'attr', 'dest', 'repo', 'branch'],
     boolean: ['verbose', 'version', 'help', 'open'],
     alias: {
       v: 'version',
       p: 'port',
       h: 'host',
-      o: 'open'
+      o: 'open',
+      a: 'attr',
+      d: 'dest',
+      r: 'repo',
+      b: 'branch'
     }
   });
 
+  const pkg = await getPackageJson();
+  updateNotifier({ pkg: pkg as Package }).notify();
+
   if (options.version) {
-    const pkg = await getPackageJson();
     console.info(pkg.version);
     return;
   }
@@ -62,6 +75,27 @@ export async function run(args: string[]) {
         host: options.host as string,
         open: Boolean(options.open),
         verbose: Boolean(options.verbose)
+      });
+      break;
+    }
+
+    case 'c':
+    case 'convert': {
+      await convert({
+        file: parameters[0],
+        attributes: options.attr as string,
+        destination: options.dest as string,
+        verbose: Boolean(options.verbose)
+      });
+      break;
+    }
+
+    case 'l':
+    case 'link': {
+      await link({
+        file: parameters[0],
+        repo: options.repo as string,
+        branch: options.branch as string
       });
       break;
     }