+++ packages/cli/src/constants.ts
@@ -1,3 +1,6 @@
 export const websitePath = 'dist/website';
 export const templatePath = 'dist/template/workshop';
+export const workshopsFolder = 'workshops/';
 export const defaultWorkshopFile = 'workshop.md';
+export const workshopBaseLink = 'https://aka.ms/ws?src=';
+export const pageBaseLink = 'https://aka.ms/ws/page?src=';